requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
h11==0.9.0
h2==3.2.0
hachoir==3.1.1
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mutagen==1.45.0
psycopg2==2.8.5
pyaes==1.6.1
pyasn1==0.4.8
pycryptodome==3.9.8
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
stagger==1.0.0
tekore==2.1.3
Telethon==1.15.0
tqdm==4.48.0
urllib3==1.25.10
waitress==1.4.4
Werkzeug==1.0.1
